Move gather_summary_observations into folder
Also renamed to more general export_observations
